moved the absorption tests for mu_s !=1 from test_absorption_mu_s.py to test_absorption.py
